### 2023-10-10 #5934 @editorialbot
[PRE REVIEW]: SpiDy.jl: open-source Julia package for the study of non-Markovian stochastic dynamics
**Submitting author:** @mekise (<a href="http://orcid.org/0000-0002-8133-1551">Stefano Scali</a>)
**Repository:** https://github.com/quantum-exeter/SpiDy.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.0
