### 2023-10-11 #5934 @mekise
4. I regenerated the pdf after I added the pdf figure in the paper subdir but it is missing in the output pdf. Is that ok?
