+++ paper/paper.md
@@ -89,7 +89,7 @@ In the next section, we show a minimal working example to run the spin dynamics 
 # Example
 
 ![**Single-spin dynamics.** Dynamics of the $x$, $y$, and 
- $z$ spin components. The components are normalized against the total spin length $S_0$ and time axes are expressed in units of the Larmor frequency $\omega_\mathrm{L}$ ($\omega_\mathrm{L} = |\mathbf{B}|$ in Eq.(1)). We show an example set of 5 stochastic trajectories of the spin dynamics (colored semi-transparent lines) together with their stochastic average (gray solid line). Note that, while we show only 5 trajectories for clarity, the average dynamics is obtained from 10000 trajectories. We also represent the range of one standard deviation from the average dynamics (gray-shaded area). In the inset, we show the convergence of the same dynamics towards the steady state at longer times. This example is obtained using the Lorentzian parameters "set 1" found in Ref. [@Anders_2022]. The code used to generate the stochastic trajectories is shown in the text. \label{fig:dynamics}](fig/dynamics.pdf){ width=100% }
+ $z$ spin components. The components are normalized against the total spin length $S_0$ and time axes are expressed in units of the Larmor frequency $\omega_\mathrm{L}$ ($\omega_\mathrm{L} = |\mathbf{B}|$ in Eq.(1)). We show an example set of 5 stochastic trajectories of the spin dynamics (colored semi-transparent lines) together with their stochastic average (gray solid line). Note that, while we show only 5 trajectories for clarity, the average dynamics is obtained from 10000 trajectories. We also represent the range of one standard deviation from the average dynamics (gray-shaded area). In the inset, we show the convergence of the same dynamics towards the steady state at longer times. This example is obtained using the Lorentzian parameters "set 1" found in Ref. [@Anders_2022]. The code used to generate the stochastic trajectories is shown in the text. \label{fig:spin_dynamics}](fig/spin_dynamics.pdf){ width=100% }
 
 Now, we show an example code to generate a run of SpiDy.jl for a single spin interacting with an environment. Given the stochastic nature of the problem solved, we will be dealing with different dynamical trajectories. In the following code, we show the parameters needed to build a subset of these trajectories, solutions to the stochastic differential equations of Eq.(1). Then, we plot a single one of them as an example. The entire code is commented throughout for a better understanding of the single elements of the run. We show the results of the dynamics averaged over a larger set (10000) of trajectories in \autoref{fig:dynamics}. Note that both the average and the standard deviation of the set of trajectories are not evaluated with the following code but are nonetheless represented in the figure for clarity. Further examples are available in the code repository.
 
