+++ paper/paper.md
@@ -44,7 +44,7 @@ We provide the user with Julia notebooks to guide them through the various mathe
 
 # Statement of need
 
-The problem of simulating the dynamics of interacting rotating bodies and harmonic oscillator networks in the presence of a dissipative environment can find a vast range of applications in the modeling of physical systems. This task is rendered particularly challenging when one desires to capture the non-Markovian effects that arise in the dynamics due to strong coupling with the environment. SpiDy.jl is a library that allows the user to efficiently simulate these systems to obtain both detailed dynamics and steady state properties.
+The problem of simulating the dynamics of interacting rotating bodies and harmonic oscillator networks in the presence of a dissipative environment can find a vast range of applications in the modeling of physical systems. This task is rendered particularly challenging when one desires to capture the non-Markovian effects that arise in the dynamics due to strong coupling with the environment. SpiDy.jl is a library that allows the user to efficiently simulate these systems to obtain both detailed dynamics and steady-state properties.
 
 A relevant example of the applicability of SpiDy.jl is the modeling of spins at low temperatures and at short timescales, which is a fundamental task to address many open questions in the field of magnetism and magnetic material modeling [@halilov98].
 State-of-the-art tools such as those developed for atomistic spin dynamics simulations are based on solving the Landau–Lifshitz–Gilbert (LLG) equation [@evans2014].
@@ -65,7 +65,7 @@ $$
 \bigg[\sum_{m\neq n}J_{n,m}\mathbf{S}_m(t) + \mathbf{B} + \mathbf{b}_n(t) + \int_{t_0}^t\mathrm{d}t^{\prime} K_n(t-t^{\prime}) \mathbf{S}_n(t^{\prime})\bigg], \quad (1)
 $$
 where $\mathbf{S}_n(t)$ represents the $n$-th spin vector, the interaction matrix $J_{n,m}$ sets the interaction strength between the $n$-th and $m$-th spins, $\mathbf{B}$ is the external field, which determines the natural precession direction and frequency of the spins in the absence of interaction, and $b_n(t)$ is the time-dependent stochastic field induced by the environment. Finally, the last integral term in Eq.(1) gives the spin dissipation due to the environment, including non-Markovian effects accounted for by the memory kernel matrix $K_n(t)$.
-Here, we allow each spin to interact with three independent sources of noise, so that in general $K_n(t) = C_n k_n(t)$, where $k_n(t)$ is a time dependent function and $C_n$ is a $3\times3$ matrix that determines how each of the $n$-th spin components couples to each of the three noise sources.
+Here, we allow each spin to interact with three independent sources of noise so that in general $K_n(t) = C_n k_n(t)$, where $k_n(t)$ is a time-dependent function and $C_n$ is a $3\times3$ matrix that determines how each of the $n$-th spin components couples to each of the three noise sources.
 To efficiently simulate the non-Markovian effects, we follow the methods explained in [@Anders_2022] and restrict ourselves to the case where the memory kernel $k(t)$ comes from a Lorentzian spectral density of the bath $\mathcal{J}(\omega) = \alpha\Gamma/((\omega_0^2 - \omega^2)^2 + \Gamma^2\omega^2)$ with peak frequency $\omega_0$, peak width $\Gamma$ and amplitude $\alpha$, so that $k(t) = \Theta(t)\,\alpha\,e^{-\Gamma t/2}\sin(\omega_1 t)/\omega_1$, where $\omega_1^2 = \omega_0^2 - \Gamma^2/4$.
[…]
