### 2023-10-20 #5934 @kyleniemeyer
Hi @mekise, to clarify: is `spectraldensities.jl` meant to be reviewed alongside this package? Because currently, the `SpiDy.jl` package alone is on the small side of what we would consider.
