### 2023-10-20 #5934 @mekise
Hi @kyleniemeyer, yes, that is correct. We branched with the package `SpectralDensities.jl` before we released `SpiDy.jl` on the Julia registry to keep things a bit more clear and modulable.
