### 2023-10-27 1f1a973
Add Chebyshev approximation functions to interp SD
