### 2023-10-27 7fe7a0b
Rename Chebysheb functions + add module entries
