+++ src/ChebyshevApprox.jl
@@ -1,4 +1,4 @@
-function cheb_row(n::Int)
+function chebrow(n::Int)
     coeffs = zeros(Int, n)
     if n==1
         coeffs[1] += 1
@@ -7,19 +7,19 @@ function cheb_row(n::Int)
         coeffs[2] += 1
         return coeffs
     end
-    coeffs[2:end] += 2*cheb_row(n-1)
-    coeffs[1:end-2] -= cheb_row(n-2)
+    coeffs[2:end] += 2*chebrow(n-1)
+    coeffs[1:end-2] -= chebrow(n-2)
     return coeffs
 end
 
-function cheb_matrix(n::Int)
+function chebmatrix(n::Int)
     matrix = zeros(Int, (n, n))
     for i in 1:n
-        matrix[i, :] = [cheb_row(i); zeros(Int, n-i)]
+        matrix[i, :] = [chebrow(i); zeros(Int, n-i)]
     end
     return matrix
 end
 
-function poly_coeffs(n, cheb_coeffs)
-    return cheb_matrix(n)' * cheb_coeffs
+function polycoeffs(n, chebcoeffs)
+    return chebmatrix(n)' * chebcoeffs
 end
