### 2023-11-06 8e44793
Change diffeqsolver to handle lists of baths
