### 2023-11-08 43c9b33
Modified spin dynamics code to allow for arbitrary number of baths
Each bath has its own Lorentzian spectral density, bfield, coupling
matrix, and a vector specifying which spins couple to the bath.
