### 2023-11-08 918a1e5
Improved readbility of differential equation code
Changes are only in readability by using more @view arrays.
Outside of the syntactic sugar, code should be exactly identical
to before this commit.
