+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"couplingfunctions/#Spin-spin-interactions","page":"Spin-spin interactions","title":"Spin-spin interactions","text":"","category":"section"},{"location":"couplingfunctions/","page":"Spin-spin interactions","title":"Spin-spin interactions","text":"Nchain\nNNlattice","category":"page"},{"location":"couplingfunctions/#SpiDy.Nchain","page":"Spin-spin interactions","title":"SpiDy.Nchain","text":"Nchain(N, J0; boundary=nothing)\n\nCreate the spin-spin coupling matrix of a 1D chain of N sites with nearest-neighbour interactions of strength set by J0.\n\nArguments\n\nN: The number of sites in the chain.\nJ0: The nearest-neighbour coupling strength.\nboundary=nothing: (Optional) Specifies the boundary condition of the chain. Default is nothing which corresponds to a chain open at the edges. Use :periodic for periodic boundary conditions.\n\nReturns\n\nAn N×N array representing the coupling matrix.\n\nExamples\n\njulia> J = Nchain(4, 2.0)\n4×4 Matrix{Float64}:\n 0.0  2.0  0.0  0.0\n 2.0  0.0  2.0  0.0\n 0.0  2.0  0.0  2.0\n 0.0  0.0  2.0  0.0\n\njulia> J_periodic = Nchain(4, 2.0; boundary=:periodic)\n4×4 Matrix{Float64}:\n 0.0  2.0  0.0  2.0\n 2.0  0.0  2.0  0.0\n 0.0  2.0  0.0  2.0\n 2.0  0.0  2.0  0.0\n\n\n\n\n\n","category":"function"},{"location":"couplingfunctions/#SpiDy.NNlattice","page":"Spin-spin interactions","title":"SpiDy.NNlattice","text":"NNlattice(N, Jh, Jv; boundary=nothing)\n\nCreate the spin-spin coupling matrix of a NxN 2D lattice with nearest-neighbour interactions and specified horizontal and vertical coupling strengths, Jh and Jv respectively.\n\nArguments\n\nN: The size of the lattice (N x N).\nJh: The horizontal nearest-neighbour coupling strength.\nJv: The vertical nearest-neighbour coupling strength.\nboundary=nothing: (Optional) Specifies the boundary condition of the lattice. Default is nothing which corresponds to open edges. Use :periodic for periodic boundary condition.\n\nReturns\n\nAn N^2×N^2 array representing the coupling matrix of the lattice.\n\nExamples\n\njulia> J = NNlattice(3, 2.0, 1.0)\n9×9 LinearAlgebra.Symmetric{Float64, Matrix{Float64}}:\n 0.0  2.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0\n 2.0  0.0  2.0  0.0  1.0  0.0  0.0  0.0  0.0\n 0.0  2.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0\n 1.0  0.0  0.0  0.0  2.0  0.0  1.0  0.0  0.0\n 0.0  1.0  0.0  2.0  0.0  2.0  0.0  1.0  0.0\n 0.0  0.0  1.0  0.0  2.0  0.0  0.0  0.0  1.0\n 0.0  0.0  0.0  1.0  0.0  0.0  0.0  2
[…]
