### 2023-11-08 5fbe057
Fix typing for local bath only case
