### 2023-11-08 057f1bc
Fix tests for multi-bath diffeqsolver
