### 2023-11-08 9166555
Fix type signature in diffeqsolver
