### 2023-12-02 97d8c54
First version dynamics inversepolykernelSD multispin multibath
