+++ src/Dynamics.jl
@@ -26,13 +26,19 @@ Note: The [`LorentzianSD`](https://quantum-exeter.github.io/SpectralDensities.jl
 # Returns
 An [`ODESolution`](https://docs.sciml.ai/DiffEqDocs/stable/basics/solution/) struct from `DifferentialEquations.jl` containing the solution of the equations of motion.
 """
-function diffeqsolver(s0, tspan, Jlist::Vector{LorentzianSD}, bfield, bcoupling::Vector{Vector{T}} where {T<:Real}, matrix::Vector{TT} where {TT<:Coupling}; JH=zero(I), S0=1/2, Bext=[0, 0, 1], saveat=[], projection=true, alg=Tsit5(), atol=1e-3, rtol=1e-3)
-    N = div(length(s0), 3)
-    M = length(Jlist)
-    u0 = [s0; zeros(6*M)]
+function diffeqsolver(s0, tspan, Jlist::Vector{InversePolyKernelSD}, bfield, bcoupling::Vector{Vector{T}} where {T<:Real}, matrix::Vector{TT} where {TT<:Coupling}; JH=zero(I), S0=1/2, Bext=[0, 0, 1], saveat=[], projection=true, alg=Tsit5(), atol=1e-3, rtol=1e-3)
     if length(Jlist) != length(matrix) || length(Jlist) != length(bcoupling)
         throw(DimensionMismatch("The dimension of Jlist, bcoupling, and matrix must match."))
     end
+    N = div(length(s0), 3)
+    M = length(Jlist)
+    D = maximum([J.deg for J in Jlist])
+    coeffs = zeros(Float64, D+1, M) # NB: the length of the coeffs vector is D+1 (degree + zero coeff)
+    for (idx, J) in enumerate(Jlist)
+        coeffs[1:J.deg, idx] = [(-1)^(k÷2)*J.coeffs[k]/J.coeffs[J.deg+1] for k in 1:J.deg]
+        coeffs[J.deg+1, idx] = (-1)^(J.deg+1)/J.coeffs[J.deg+1]
+    end
+    u0 = [s0; zeros(3*M*D)]
     Cω2 = []
     b = []
     for i in 1:M
@@ -43,33 +49,44 @@ function diffeqsolver(s0, tspan, Jlist::Vector{LorentzianSD}, bfield, bcoupling:
         Cω2v = get_tmp(Cω2v, u)
         Beff = get_tmp(Beff, u)
         s = @view u[1:3*N]
-        v = @view u[1+3*N:3*N+3*M]
-        w = @view u[1+3*N+3*M:3*N+6*M]
+        v = @view u[1+3*N:3*N+3*M*D]
+        ds = @view du[1:3*N]
+        dv = @view du[1+3*N:3*N+3*M*D]
         for i in 1:N
-            Beff[i, :] .= Bext 
+            Beff[i, :] .= Bext
             for j in 1:M
-                Beff[i, :] .+= bcoupling[j][i]*(b[j](t) + mul!(Cω2v, Cω2[j], v[1+(j-1)*3:3+(j-1)*3]))
+                Beff[i, :] .+= bcoupling[j][i] * (b[j](t) + mul!(Cω2v, Cω2[j], v[1+(j-1)*3*D:3+(j-1)*3*D]))
             end
         end
         for i in 1:N
-            du[1+(i-1)*3] = -(s[2+(i-1)*3]*Beff[i,3]-s[3+(i-1)*3]*Beff[i,2])
-            du[2+(i-1)*3] = -(s[3+(i-1)*3]*Beff[i,1]-s[1+(i-1)*3]*Beff[i,3])
[…]
