### 2023-12-06 e92272e
CompatHelper: add new compat entry for Statistics at version 1, (keep existing compat)
