+++ Project.toml
@@ -21,6 +21,7 @@ FFTW = "1"
 Interpolations = "0.14"
 PreallocationTools = "0.4"
 SpectralDensities = "0.2"
+Statistics = "1"
 julia = "1.7"
 
 [extras]
