+++ Project.toml
@@ -18,7 +18,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 DifferentialEquations = "7"
 Distributions = "0.25"
 FFTW = "1"
-Interpolations = "0.14"
+Interpolations = "0.14, 0.15"
 PreallocationTools = "0.4"
 SpectralDensities = "0.2"
 julia = "1.7"
