### 2024-01-10 3c0042b
Bumped version number to 0.2
