### 2024-01-23 #6263 @editorialbot
[REVIEW]: SpiDy.jl: open-source Julia package for the study of non-Markovian stochastic dynamics
**Submitting author:** @mekise (<a href="http://orcid.org/0000-0002-8133-1551">Stefano Scali</a>)
**Repository:** https://github.com/quantum-exeter/SpiDy.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
