### 2024-01-23 #6263 @matthewfeickert
@mekise @Datseris, sorry I missed this during the pre-review, but I now see that you're both employed at the University of Exeter which might be a [Conflict of Interest violation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). It looks like you work in seperate departments (Physics vs. Mathematics and Statistics), but can you provide context on if you have any overlap at the University?
