### 2024-01-23 #6263 @matthewfeickert
> No overlap at all.

@mekise @Datseris okay great. I think we can move forward with the review then with the conflict being recorded and then waived, as this having no overlap means you fall into the example of both being "[employed by the same very large organization but in different units without any knowledge of each other.](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)" :+1:
