### 2024-01-27 #6263 @Datseris
Hi @matthewfeickert ,

before investing more time in the review, we need to clarify something regarding the scholarly effort point. I read the JOSS guidliness for substantial scholarly effort as I was ticking the relevant checkmark: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort . But then I decided to check [the source code](https://github.com/quantum-exeter/SpiDy.jl/blob/main/src) to confirm my tick. The source code is much less than 1000 loc, and by my count it just reaches 300 (I've excluded documentation in counting source loc) which according to the current guidelines would not be enough to satisfy the requirements.

I have not yet took a detailed look at the paper itself. The repository source code however appears to be a structuring of logical containers for running stochastic differential equations. It sets up the equations and the noise term and then runs DifferentialEquations.jl (the julia package for running SDEs) for the user: https://github.com/quantum-exeter/SpiDy.jl/blob/3c0042b87381707a43dcae39c65eed823dc46808/src/Dynamics.jl#L28 . `diffeqsolver` which is the main export of the software is doing what I described above.

Please advise how this relates to the guidelines before we invest more time into the review!

all the best,
George
