### 2024-01-28 #6263 @mekise
Hi @Datseris and @matthewfeickert , just wanted to mention that this was already been raised and resolved in pre-review (see https://github.com/openjournals/joss-reviews/issues/5934#issuecomment-1799230068). Hope this helps.
