### 2024-02-02 quantum-exeter/SpiDy.jl#16 @mdavezac
Community guidelines
Submissions to JOSS require some form of community guidelines for contributing to the code.
In general, this takes the form of a `contributing` file as explained [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#adding-a-contributing-file), and [here](https://github.com/czbiohub-sf/cerebra/blob/master/docs/CONTRIBUTING.md) is an example of the content of the file.

SpectralDensities.jl will also need one since it is covered in the JOSS publication.

If you want to go further (not required for JOSS) and add a code of conduct, PR and issue templates, github's [community profile page](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/about-community-profiles-for-public-repositories) is not bad place to start
