### 2024-02-02 quantum-exeter/SpiDy.jl#16 @mekise
Thanks, I will go through the community standards soon adding the relative files.
