### 2024-02-02 quantum-exeter/SpiDy.jl#17 @mdavezac
Testing
JOSS requires sufficient testing for correctness.
It looks like SpectralDensities.jl tests for quite a few things, but SpiDy.jl only checks the code runs, not that the output is sensible.

It would be good to add a few tests that check output values, preferably some physically significant result, but reproducibility against a specific random seed is also okay.
