### 2024-02-17 #6263 @Datseris
Hi, I am submitting my review:

I ticked off most of my review checkboxes. However, I do have some standing concerns:

1. The tests tickbox I left open as the main repo does not appear to have adequate tests. It is only checking that the ODE Solving yielded a "success" code. But this just means that solutions did not diverge to infinity. There are no checks whatsoever regarding whether the solutions are physically correct or whether they reflect the properties expected by the physical system they emulate. The tests must be made more extensive and test for the accuracy of setting up the problem and the resulting physical solution.
2. I think the name of the software can be improved. The software description reads: "Non-Markovian stochastic dynamics for spin and harmonic oscillators. " and the paper summary introduces it as a software to "solve the non-Markovian stochastic dynamics of interacting classical spin vectors and harmonic oscillator networks in contact with a dissipative environment". There is simply no way one can associate the abbreviated name "SpiDy" with this very specific target application. Even "Spin Dynamics" appears to not be an accurate name, as the software is described as much more specific than treating any kind of spin dynamics (such as Ising models and quantum statistical mechanics type of problems). Additionally half the name of the software is supposed to reflect "Dynamics", which by itself is not only a rather general word, but also already heavily occupied by many other julia packages as well as entire organizations (JuliaDynamics). I think it is to the benefit of the authors to change the name to something that will help people not familiar yet with the software understand whether this software is for them or not at a first glance.
3. The performance discussion in the paper is not adequate I believe, as the only statement I could find was "The package is written in pure Julia to take advantage of the language performance". Just using Julia is not enough to bring optimal performance versus e.g., Python. Did the authors take any steps/considerations when creating the ODE problem to be passed to DifferentialEquations.jl to maximize performance? Did they utilize caches, did they parallelize anything via threads or other means, etc.
4. I did not find community guidelines in the docs.
