### 2024-02-19 #6263 @mdavezac
I've raised two issues in the main repo regarding [testing](https://github.com/quantum-exeter/SpiDy.jl/issues/17) and [community guidelines](https://github.com/quantum-exeter/SpiDy.jl/issues/16). The latter is also an issue with SpectralDensities.jl.  Except for these outstanding issues, I believe SpiDy.jl and SpectralDensities.jl satisfy all requirements for publication in JOSS.
