### 2024-03-01 53b2e28
Network added for oscillators
