+++ src/Dynamics.jl
@@ -124,23 +124,23 @@ function diffeqsolver(s0, tspan, J::LorentzianSD, bfield, matrix::Coupling; JH=z
 end
 
 """
-    function diffeqsolver(s0, tspan, J::LorentzianSD, Jshared::LorentzianSD, bfields, bfieldshared, matrix::Coupling; JH=zero(I), S0=1/2, Bext=[0, 0, 1], saveat=[], projection=true, alg=Tsit5(), atol=1e-3, rtol=1e-3)
+    function diffeqsolver(x0, p0, tspan, Jlist::Vector{LorentzianSD}, bfield, bcoupling::Vector{<:AbstractArray{T,1}} where {T<:Real}, matrix::Vector{TT} where {TT<:Coupling}; JH=zero(I),  Ω=1.0, counter_term=true, saveat=[], alg=Tsit5(), atol=1e-3, rtol=1e-3)
 
-Solves the dynamics of a system of ineracting spins under the influence of *either* local (unique to each spin)
-or global (shared by all spins) stochastic noise from the environment.
+Solves the dynamics of a system of ineracting oscillators under the influence of *both* local (unique to each oscillator)
+and global (shared by all oscillators) stochastic noise from the environment.
 
 # Arguments
 - `x0`: Array of length `N` specifying the initial position of the `N` oscillators.
 - `p0`: Array of length `N` specifying the initial momentum of the `N` oscillators.
 - `tspan`: The time span to solve the equations over, specified as a tuple `(tstart, tend)`.
-- `J::LorentzianSD`: The spectral density of the noise acting globally on the harmonic oscillators.
-- `bfields`: A tuple of functions `Tuple{Function, Function, Function}` representing the time series of the global stochastic field shared by all the harmonic oscillators.
-- `matrix::Coupling`: The harmonic oscillators-environment coupling matrix.
+- `Jlist::Vector{LorentzianSD}`: The list of spectral densities of the environment(s).
+- `bfield`: An array of tuples of functions `Array{Tuple{Function, Function, Function}}` representing the time series of the noise for each environment.
+- `bcoupling::Vector{<:AbstractArray{T,1}} T <: Real`: A vector (of length the number of baths) of vectors (of length the number of oscillators), specifying if each bath couples to each oscillator.
+- `matrix::Vector{TT} TT <: Coupling`: A vector of `Coupling` structs specifying the oscillator-environment coupling matrix for each environment.
 - `JH=zero(I)`: (Optional) The oscillator-oscillator coupling matrix. Default is zero matrix (i.e. non-interacting oscillators).
-- `Ω=1`: (Optional) The natural angular frequency of the harmonic oscillators (currently the same for all). Default is 1.
+- `Ω=1.0`: (Optional) The natural angular frequency of the harmonic oscillators (currently the same for all). Default is 1.
 - `counter_term=true`: (Optional) Whether to include the counter-term or not. Default is true.
 - `saveat=[]`: (Optional) An array of time points where the solution should be saved. Default is empty, which saves the solution at the time steps chosen by the integration algorithm.
[…]
