### 2024-03-01 ab39777
Added option to output auxillary memory fields
