### 2024-03-01 5bf32ab
Fixed indexing for spin case
