+++ Project.toml
@@ -1,7 +1,7 @@
 name = "SpiDy"
 uuid = "a4849050-cccd-45e7-a409-2b114d5ccb2f"
 authors = ["Stefano Scali <scali.stefano@gmail.com>", "Federico Cerisola <federico@cerisola.net>"]
-version = "0.2.0"
+version = "0.2.1"
 
 [deps]
 DifferentialEquations = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
