### 2024-04-11 03822d3
Create CODE_OF_CONDUCT.md
