+++ SECURITY.md
@@ -1,0 +1,8 @@
+# Security Policy
+
+Please see "Reporting Bugs" in [CONTRIBUTING.md](https://github.com/quantum-exeter/SpiDy.jl/blob/main/CONTRIBUTING.md) to report to us any security issue you
+find in SpiDy.jl.
+
+⚠️ DO NOT use the issue tracker or discuss vulnerabilities as it will cause more damage than help.
+
+Please note that as a non-commercial open-source project, we are not able to pay bounties.
