+++ .github/pull_request_template.md
@@ -1,0 +1,7 @@
+## Pull request
+
+To open a pull request, please do the following.
+
+- Reference to a related issue addressed by this pull request
+- Describe the changes proposed in the pull request
+- Mention the person responsible for reviewing the proposed changes
