### 2024-04-11 quantum-exeter/SpiDy.jl#16 @mekise
All community guidelines documents added with commits 03822d316e5e84e9cd8811f9889d246d35b62d77 a204dc5387340a1616ed9a5069c996f96e3b2eb5 cb3c70afbf2fde344af53f5cbe73f1ebfb2ff149 c8ed2566d4b391dc6c99445539463a1686998117 7c1c0ddc0ea0c52c8cab4f231fdd029172c938c2
