### 2024-04-11 5f3c784
Added unit tests for dynamics
Added a test at T=0 compared with analytics and a test of stochastic
dynamics with a fixed random seed. Fixes issue #17.
