### 2024-04-11 quantum-exeter/SpiDy.jl#17 @cerisola
Unit tests for the dynamics have now been added in 5f3c784b38366250f63eca5c9914081adb710e80:

1. We test the dynamics at $T=0$ with classical noise against an analytical solution that is known for this case.
2. We test the dynamics with non-trivial stochastic noise using a fixed random seed.

I think this addresses the issue raised so I'll go ahead and close the it. Feel free to comment and reopen if you think further testing should be included.
