### 2024-04-11 #6263 @mekise
Hi @mdavezac

Thank you for your assessment. We are pleased to see that SpiDy.jl satisfies the requirements for publication.
1. We completed and added all the documents needed to meet the community guidelines. We did this in both SpiDy.jl and SpectralDensities.jl. We addressed and closed the relevant issues.
