### 2024-04-11 #6263 @mekise
> Hi, I am submitting my review:
>
> I ticked off most of my review checkboxes. However, I do have some standing concerns:
>
> 1. The tests tickbox I left open as the main repo does not appear to have adequate tests. It is only checking that the ODE Solving yielded a "success" code. But this just means that solutions did not diverge to infinity. There are no checks whatsoever regarding whether the solutions are physically correct or whether they reflect the properties expected by the physical system they emulate. The tests must be made more extensive and test for the accuracy of setting up the problem and the resulting physical solution.
> 2. I think the name of the software can be improved. The software description reads: "Non-Markovian stochastic dynamics for spin and harmonic oscillators. " and the paper summary introduces it as a software to "solve the non-Markovian stochastic dynamics of interacting classical spin vectors and harmonic oscillator networks in contact with a dissipative environment". There is simply no way one can associate the abbreviated name "SpiDy" with this very specific target application. Even "SpinDynamics.jl" appears to not be an accurate name, as the software is described as much more specific than treating any kind of spin dynamics (such as Ising models and quantum statistical mechanics type of problems). I think it is to the benefit of the authors to change the name to something that will help people not familiar yet with the software understand whether this software is for them or not at a first glance.
> 3. The performance discussion in the paper is not adequate I believe, as the only statement I could find was "The package is written in pure Julia to take advantage of the language performance". Just using Julia is not enough to bring optimal performance versus e.g., Python. Did the authors take any steps/considerations when creating the ODE problem to be passed to DifferentialEquations.jl to maximize performance? Did they utilize caches, did they parallelize anything via threads or other means, etc.
> 4. I did not find community guidelines in the docs.

Hi @Datseris

1. We agree that testing of physically correct results are needed. We now added such testing. Given the stochastic nature of the problem solved, we approached the problem from two different perspectives. First, we added a test at zero temperature of the solution of the set of differential equations. This is done to remove the stochastic component from the system and it is checked against the analytical solution that we derive in this regime. Second, we added a fixed-seed test. Such test is added to verify consistency of the solution when the stochastic component is present. We check the result against stochastic results found in the literature.

2. We understand the concerns of the reviewer about the name of the package. We note that such naming is derived a few years ago by the initial mission of the package. The mission was indeed to improve the classical simulation in the atomistic "Spin Dynamics" (ASD) community (see for example https://doi.org/10.1063/1.4930971). Hence, the name SpiDy. In time, we improved and updated the package to make it deal with different spectral densities, different noise, multiple spins, and even added a variant to solve systems of coupled harmonic oscillators. But, while the range of possible applications has expanded in time, the main applicative use of SpiDy.jl remains the atomistic spin dynamics community. For this reason, we believe the the name is still relevant and we would like to keep it as is.

3. We agree that the sentence of performance indicated by the reviewer is not backed up by any numerical comparison or asymptotics. What we wanted to express was intrinsically related to the use of the DifferentialEquations.jl package in Julia. Such package is notoriously an efficient package for solving sets of differential equations, allowing for vast method customizations at the same time. We followed the package performance guidelines during our implementation reducing evaluation redundancy, using callbacks, and preallocation (get_tmp to make it type-stable and non-allocating). To make the text more precise and consistent we changed the sentence to "The package is written in pure Julia and we take advantage of the efficient DifferentialEquations.jl package by reducing evaluation redundancy, using callbacks, and pre-allocations".

4. Thank you for reminding us. We completed and added all the documents needed to meet the community guidelines.
