### 2024-04-11 8e7e98c
Update paper.md for JOSS review
