+++ paper/paper.md
@@ -52,8 +52,7 @@ Despite their massive success, these tools run into shortcomings in accurately m
 Recent work has focused on developing a comprehensive quantum-thermodynamically consistent framework suitable to model the dynamics of spins in magnetic materials while addressing these shortcomings [@Anders_2022]. This framework includes strong coupling effects to the environment such as non-Markovian memory, colored noise, and quantum-like fluctuations.
 At its core, SpiDy.jl implements the theoretical framework introduced in [@Anders_2022], allowing for the study of environment memory effects and anisotropic system-environment coupling.
 SpiDy.jl can be readily adopted for atomistic spin dynamics simulations [@evans2014; @Barker_2019], ultrafast magnetism [@Beaurepaire_1996], and ferromagnetic and semiconductive systems exhibiting anisotropic damping [@Chen_2018].
-A further set of applications stems from the extension of SpiDy.jl to handle the non-Markovian stochastic dynamics of harmonic oscillators. This model will be of interest in the field of quantum thermodynamics where harmonic oscillators play a key role in modeling open quantum systems.
-The package is written in pure Julia to take advantage of the language performance.
+A further set of applications stems from the extension of SpiDy.jl to handle the non-Markovian stochastic dynamics of harmonic oscillators. This model will be of interest in the field of quantum thermodynamics where harmonic oscillators play a key role in modeling open quantum systems. The package is written in pure Julia and we take advantage of the efficient DifferentialEquations.jl package by reducing evaluation redundancy, using callbacks, and pre-allocations.
 
 The software package has seen a wide range of applications to date. Firstly, the convenience of three independent environments in SpiDy.jl finds application in the microscopic modeling of spins affected by noise due to vibrations of the material lattice [@Anders_2022]. SpiDy.jl also found application in the demonstration of the quantum-to-classical correspondence at all coupling strengths between a spin and an external environment [@cerisola2022quantumclassical]. Here, the temperature dependence of the spin steady-state magnetization obtained with SpiDy.jl is successfully compared with the classical mean force state of the system. In Ref. [@hartmann2023anisotropic], the authors take advantage of the customizable coupling tensor in SpiDy.jl to explore the anisotropic effects of the environment on the system. In Ref. [@berritta2023accounting], SpiDy.jl is used as a sub-routine to build quantum-improved atomistic spin dynamics simulations. In the paper, the authors take advantage of the customizable power spectrum to implement ad-hoc simulations matching known experimental results. Lastly, with an eye to the harmonic oscillator side, SpiDy.jl is used to match the quantum harmonic oscillator dynamics with its stochastic counterpart [@glatthard2023harmonic]. Here, the authors exploit the recent implementation of harmonic oscillator dynamics.
 
