### 2024-05-21 #6263 @matthewfeickert
> MISSING DOIs
>
> - No DOI given, and none found for title: In preparation

@mekise Do you know if [the reference](https://github.com/quantum-exeter/SpiDy.jl/blob/8e7e98c5b47850e01c48625807ce810bf6d3a509/paper/paper.bib#L59C1-L63C2)

```bib
@misc{glatthard2023harmonic,
    author = {Jonas Glatthard and Charlie Hogg and Federico Cerisola and Janet Anders},
    title = {In preparation},
    year = {2023}
}
```

is still in preparation, or do you know if an arXiv version or a publication is available now? If not, that's okay, but if it there is something with a title that is citable that is preferable.
