### 2024-05-21 #6263 @mekise
@matthewfeickert The reference is yet to come out as a preprint but in the final rounds or revision. The entry can be updated as:

@misc{hogg2024stochastic,
    author = {Charlie Hogg and Jonas Glatthard and Federico Cerisola and Janet Anders},
    title = {Stochastic simulation of dissipative quantum oscillators},
    journal = {In preparation},
    year = {2024}
}

Let me know if I need to update it or will postpone it to a full set of changes when the editorial phase is complete.
