### 2024-05-21 #6263 @mekise
Also, I have noticed that there are another couple of DOIs I can update from the arxiv version to the published version. I will keep them in mind when I will change the refs.
