### 2024-05-22 quantum-exeter/SpiDy.jl#23 @matthewfeickert
[JOSS Review]: Suggested paper revisions
These are some comments to help the paper component of https://github.com/openjournals/joss-reviews/issues/6263 be more compliant with a JOSS paper.

* [x] Simon Horsley's ORCID of https://orcid.org/0000-0001-7242-7941 and Janet Anders's ORCID of https://orcid.org/0000-0002-9791-0363 seem to be empty. Having information in the ORCID is not required for publication, so this is fine, but if this was not intended they should double check their accounts.

https://github.com/quantum-exeter/SpiDy.jl/blob/8e7e98c5b47850e01c48625807ce810bf6d3a509/paper/paper.md?plain=1#L17-L21

* [x] In

https://github.com/quantum-exeter/SpiDy.jl/blob/8e7e98c5b47850e01c48625807ce810bf6d3a509/paper/paper.md?plain=1#L55

specifically

> The package is written in pure Julia and we take advantage of the efficient DifferentialEquations.jl package...

please cite [`DifferentialEquations.jl`](https://github.com/SciML/DifferentialEquations.jl). They [provide a `CITATION.bib`](https://github.com/SciML/DifferentialEquations.jl/blob/ad7b74d096f2883c5686c48a11f4edd94eef644e/CITATION.bib):

```bib
@article{DifferentialEquations.jl-2017,
 author = {Rackauckas, Christopher and Nie, Qing},
 doi = {10.5334/jors.151},
 journal = {The Journal of Open Research Software},
 keywords = {Applied Mathematics},
 note = {Exported from https://app.dimensions.ai on 2019/05/05},
 number = {1},
 pages = {},
 title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
 url = {https://app.dimensions.ai/details/publication/pub.1085583166 and http://openresearchsoftware.metajnl.com/articles/10.5334/jors.151/galley/245/download/},
 volume = {5},
 year = {2017}
}
```

* [x] All of the ["Example" section](https://github.com/quantum-exeter/SpiDy.jl/blob/8e7e98c5b47850e01c48625807ce810bf6d3a509/paper/paper.md?plain=1#L88-L143) should be removed. While I understand the motivation to include it, in the [What should my paper contain?](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) JOSS docs examples are not a requested paper section and we recommend adding all such useful examples to the documentation, if they are not already there.

* [x] While the ["Overview" section](https://github.com/quantum-exeter/SpiDy.jl/blob/8e7e98c5b47850e01c48625807ce810bf6d3a509/paper/paper.md?plain=1#L59-L86) is not requested, if it is important to you to keep it you may, but it will need to be reduced in scope. As noted in the docs, the paper should include

> A summary describing the high-level functionality and purpose of the software for a diverse, _non-specialist_ audience.

where "non-specialist" means someone outside of your field. You could probably reduce this entire section down to a modified version of

https://github.com/quantum-exeter/SpiDy.jl/blob/8e7e98c5b47850e01c48625807ce810bf6d3a509/paper/paper.md?plain=1#L78-L84

though I appreciate that you might feel the need to give a bit more supporting detail, which is fine.

* [x] We already discussed this in https://github.com/openjournals/joss-reviews/issues/6263#issuecomment-2121972051, but just to add it to the checklist, the `glatthard2023harmonic`/`hogg2024stochastic` reference should get updated.

Beyond that I think the paper is great and once this is done we can move forward with the versioning and archiving steps of publishing. :+1:

Please let me know of any questions or concerns that you might have, and I'll be happy to try to help address them.
