### 2024-05-22 #6263 @matthewfeickert
I've made some editorial comments on the paper draft in https://github.com/quantum-exeter/SpiDy.jl/issues/23, which I hope will be easy to implement, but please let me know of any questions or concerns that you have. :+1:
