### 2024-05-22 88b3f8b
Update paper folder and added run_example_paper to meet editor requests
