### 2024-05-22 129c82a
Update bib and md file to comply with editor suggestions
