+++ paper/paper.bib
@@ -24,14 +24,15 @@
 }
 
 @article{cerisola2022quantumclassical,
-    doi={10.48550/arXiv.2204.10874},
-    title={Quantum-classical correspondence in spin-boson equilibrium states at arbitrary coupling}, 
-    author={Federico Cerisola and Marco Berritta and Stefano Scali and Simon A. R. Horsley and James D. Cresser and Janet Anders},
-    year={2022},
-    eprint={2204.10874},
-    archivePrefix={arXiv},
-    primaryClass={quant-ph},
-    journal={arXiv preprint}
+    title={Quantum-classical correspondence in spin-boson equilibrium states at arbitrary coupling},
+    ISSN={1367-2630},
+    url={http://dx.doi.org/10.1088/1367-2630/ad4818},
+    DOI={10.1088/1367-2630/ad4818},
+    journal={New Journal of Physics},
+    publisher={IOP Publishing},
+    author={Cerisola, Federico and Berritta, Marco and Scali, Stefano and Horsley, Simon A R and Cresser, James D and Anders, Janet},
+    year={2024},
+    month=may 
 }
 
 @article{berritta2023accounting,
@@ -56,10 +57,11 @@
     journal={arXiv preprint}
 }
 
-@misc{glatthard2023harmonic,
-    author = {Jonas Glatthard and Charlie Hogg and Federico Cerisola and Janet Anders},
-    title = {In preparation},
-    year = {2023}
+@misc{hogg2024stochastic,
+    author = {Charlie Hogg and Jonas Glatthard and Federico Cerisola and Janet Anders},
+    title = {Stochastic simulation of dissipative quantum oscillators},
+    journal = {In preparation},
+    year = {2024}
 }
 
 @article{evans2014,
@@ -137,4 +139,18 @@
 	author = {L. Chen and S. Mankovsky and S. Wimmer and M. A. W. Schoen and H. S. Körner and M. Kronseder and D. Schuh and D. Bougeard and H. Ebert and D. Weiss and C. H. Back},
 	title = {Emergence of anisotropic Gilbert damping in ultrathin Fe layers on {GaAs}(001)},
 	journal = {Nature Physics}
+}
+
+@article{DifferentialEquations_jl,
+        author = {Rackauckas, Christopher and Nie, Qing},
+        doi = {10.5334/jors.151},
+        journal = {The Journal of Open Research Software},
+        keywords = {Applied Mathematics},
+        note = {Exported from https://app.dimensions.ai on 2019/05/05},
+        number = {1},
+        pages = {},
+        title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+        url = {https://app.dimensions.ai/details/publication/pub.1085583166 and http://openresearchsoftware.metajnl.com/articles/10.5334/jors.151/galley/245/download/},
+        volume = {5},
+        year = {2017}
 }
