### 2024-05-22 quantum-exeter/SpiDy.jl#23 @mekise
Hi @matthewfeickert , I have address the tasks.

1. the orcids are correct even if empty, double checked.
2. added citation for the package DifferentialEquations.jl
3. removed example section and referred to a new jl file added to the example runs folder
4. drastically reduced the overview section. I left the basic equation solved by the package to showcase the different "pieces" that can be customized in the package. Removed most of the technicalities. If possible, I would like to keep a minimal overview of the maths behind for the reader to be able to then refer to the paper I cite, if needed.
5. fixed the citation and added another DOI
