### 2024-05-22 #6263 @mekise
Hi @matthewfeickert , I have regenerated the pdf and it looks ok to me. Two observations:

1. this issue is named "SpiDy.jl: open-source Julia package for the study of 2 non-Markovian stochastic dynamics" but there should be no "2", I do not know if this will become an issue in the future, but I wanted to point it out.
2. the paper is now in the main branch, is it possible to move the bot from the paper branch to main? It would be quicker for me to push in main now. Thanks
