### 2024-05-22 #6263 @matthewfeickert
Thanks @mekise! I'll review your paper edits later today once I'm free of work meetings (which won't be until the afternoon US time).

> this issue is named "SpiDy.jl: open-source Julia package for the study of 2 non-Markovian stochastic dynamics" but there should be no "2", I do not know if this will become an issue in the future, but I wanted to point it out.

Good catch. I assume this was some bug of `@editorialbot` when it created the issue from the pre-review. I've changed it, but luckily the issue title isn't used in any paper metadata.

> the paper is now in the main branch, is it possible to move the bot from the paper branch to main? It would be quicker for me to push in main now. Thanks

Yup and done. :+1:
