+++ paper/paper.md
@@ -73,6 +73,8 @@ In conclusion, SpiDy.jl implements the stochastic dynamics of coupled integro-di
 - Custom system-environment coupling tensors, allowing for isotropic or anisotropic couplings. Both amplitudes and geometry of the coupling can be specified.
 - Choice between local environments, i.e. distinct baths acting on the single sub-system, or a single common environment.
 
+We show an example run of the stochastic trajectories in the case of a single spin in \autoref{fig:dynamics}. The code used to generate this example can be found (along with comments) in the GitHub repository of SpiDy.jl.
+
 ![**Single-spin dynamics.** Dynamics of the $x$, $y$, and 
  $z$ spin components. The components are normalized against the total spin length $S_0$ and time axes are expressed in units of the Larmor frequency $\omega_\mathrm{L}$ ($\omega_\mathrm{L} = |\mathbf{B}|$ in Eq.(1)). We show an example set of 5 stochastic trajectories of the spin dynamics (colored semi-transparent lines) together with their stochastic average (gray solid line). Note that, while we show only 5 trajectories for clarity, the average dynamics is obtained from 10000 trajectories. We also represent the range of one standard deviation from the average dynamics (gray-shaded area). In the inset, we show the convergence of the same dynamics towards the steady state at longer times. This example is obtained using the Lorentzian parameters "set 1" found in Ref. [@Anders_2022]. The code used to generate the stochastic trajectories is included in the GitHub repository. \label{fig:dynamics}](spin_dynamics.pdf){ width=100% }
 
