### 2024-05-22 quantum-exeter/SpiDy.jl#23 @matthewfeickert
The draft as of https://github.com/openjournals/joss-reviews/issues/6263#issuecomment-2124932323 looks good, so closing this as resolved.
