### 2024-05-22 quantum-exeter/SpiDy.jl#24 @matthewfeickert
[JOSS Review]: Recommend attributing an author as copyright holder in the LICENSE
Editorial notes as part of https://github.com/openjournals/joss-reviews/issues/6263:

The current LICENSE file doesn't actually attribute any author with the copyright

https://github.com/quantum-exeter/SpiDy.jl/blob/72259092247ca2131496c542e1691f778e9dd9d3/LICENSE#L474

It is recommended that some author(s) be added to the LICENSE (e.g. "SpiDy.jl Developers") along with the original development year.
