### 2024-05-22 #6263 @matthewfeickert
@mekise all that's left to do now (after addressing https://github.com/quantum-exeter/SpiDy.jl/issues/24) before recommending publication is to ensure that there is a long term public archive of the code that was reviewed. In this case the last code release was [`SpiDy.jl` `v0.2.1`](https://github.com/quantum-exeter/SpiDy.jl/releases/tag/v0.2.1) though there has been [some development between then and the current `HEAD`](https://github.com/quantum-exeter/SpiDy.jl/compare/v0.2.1...72259092247ca2131496c542e1691f778e9dd9d3), so we'd suggest making a new release (as you see fit with respect to SemVer).

We'd suggest depositing the code either with [Zenodo](https://zenodo.org/) or with [figshare](https://figshare.com/) to get an archive with a DOI. If you use Zenodo there is [an (optional) GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) that can create a Zenodo archive for you anytime you make a GitHub release of your code.

[![Zenodo_DOI_guide](https://user-images.githubusercontent.com/5142394/234173148-2226325d-0a42-4526-b546-75b469d26c4c.png)](https://zenodo.org/account/settings/github/)

Once you make a new release of `SpiDy.jl` and have a DOI for the archive please just report them here and I'll have @editorialbot update the version listed in the review and add the archive.

Let me know if you have any questions. :+1:
