### 2024-05-24 quantum-exeter/SpiDy.jl#24 @mekise
Fixed with commit 11cabd47f1a80981d6031f4ec2f60eed093cfdb8
