### 2024-05-24 #6263 @mekise
Hi @matthewfeickert I have done all that you requested, new release, Zenodo archive, and DOI. Plus, fixed https://github.com/quantum-exeter/SpiDy.jl/issues/24 . Everything should be in order.

version: v1.0.0
DOI: 10.5281/zenodo.11283978
