### 2024-05-24 #6263 @matthewfeickert
@mekise Thanks for taking care of that! I'll go ahead and move forward with recommending acceptance for publication, but while I'm doing that can you please edit the metadata of https://zenodo.org/records/11283978 to

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

? We can do these things asynchronously though. :+1:
