### 2024-05-24 #6263 @matthewfeickert
@openjournals/pe-eics note https://github.com/openjournals/joss-reviews/issues/6263#issuecomment-2129969033 for archive metadata edits that need to happen.
