### 2024-05-29 #6263 @kyleniemeyer
@mekise just some small edits to the paper needed. Near the end of the statement of need setion, you have:
> In Ref. (Hartmann et al., 2023), the authors take advantage of the customizable coupling tensor in SpiDy.jl to explore the anisotropic effects of the environment on the system. In Ref. (Berritta et al., 2023)...

Could you edit the way those references are mentioned, like "Hartmann et al. (2023) take advantage..." and "Berritta et al. (2023) use SpiDy.jl as a sub-routine..."? Citing in the `@author:2001` format (no square brackets) will lead to the references appearing this way.

Please also update the mention of the Anders et al. (2002) reference in the Figure 1 caption in the same manner.
