### 2024-05-30 ee656e2
Change style of some citations in paper
