### 2024-05-30 #6263 @mekise
Hi @kyleniemeyer , I removed the square brackets as requested. Is this the right result?
